Split the files this branch pushed past the size ratchet, and clear the footguns
CI runs two checks AGENTS.md does not list among its four gates, and both failed.

check-footguns.mjs flagged two things the branch introduced. An absolute /Users/
path in hook-adapter.test.ts -- and the personal-path rule catches it, unlike
privacy-guard.test.ts, which only matches a fixed list of past contributor names;
the fixtures are now /projects/..., with no home directory at all. And a
process.env read outside the config layer in claude-interactive.ts, now threaded
as a parameter, which is the fix the rule itself names. One suppression remains,
with its reason: an unrelated pre-existing read in server.ts that came into scope
only because this branch had to touch that file.

ratchet.mjs found nine files over budget. Every one is split by responsibility,
which is what size-baseline.json says to do; the baseline is re-run and committed
tighter, never widened by hand:

  engines/omniroute-upstream.ts     the whole SSE proxy start, out of the 1859-line
                                    claude-interactive.ts
  gateway/memory-trial-route.ts     hook routing, engine-session capture and the
                                    SessionStart extra, out of api.ts
  gateway/delegation-brief.ts       the canonical Todo dossier and the review-
                                    dispatch rule, out of api.ts
  workflow-todo-runs.ts             the memory-workflow archiver, out of server.ts,
                                    in a module server.ts already imports
  shared/config-memory-trial.ts     memoryTrial validation, one function per concern
  memory-trial/claims.ts            the pure claim predicates, out of guardrails.ts
  __tests__/*-fixtures.ts           the two test harnesses both halves now share
  workflow/run-input.ts, run-button.tsx   the run-input prompt and its button

Two rules bit only after the move, because the files they came from carry ESLint
grandfather exemptions and new files do not: workflowRunOnChange was 55 lines
against a 50 limit and its inner arrow had a complexity of 22, and
validateMemoryTrial had a complexity of 23. Both are split further rather than
exempted. An imported `let` is read-only at the import site, so the engine-
availability flag the delegation suite toggles moved into a holder object.

Green now: typecheck, lint, test (997 vitest files plus 210 node tests), build,
footguns, ratchet.
