+++ packages/web/src/routes/workflow/page.tsx
@@ -1,7 +1,8 @@
 import { useCallback, useState } from "react"
-import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query"
-import { ArrowLeft, Play } from "lucide-react"
-import { Link, useNavigate, useParams, useSearchParams } from "react-router-dom"
+import { RunButton } from "./run-button"
+import { useQuery, useQueryClient } from "@tanstack/react-query"
+import { ArrowLeft } from "lucide-react"
+import { Link, useParams, useSearchParams } from "react-router-dom"
 import { PageLayout } from "@/components/page-layout"
 import {
   ApiError,
@@ -90,52 +91,6 @@ function RunsSection({ workflowId }: { workflowId: string }) {
         </div>
       )}
     </section>
-  )
-}
-
-function collectRunInput(inputs: WorkflowDefinitionWire["inputs"]): Record<string, string> | null {
-  const values: Record<string, string> = {}
-  for (const input of inputs ?? []) {
-    const value = window.prompt(
-      `${input.label}${input.description ? `\n\n${input.description}` : ""}`,
-      typeof input.default === "string" ? input.default : "",
-    )
-    if (value === null) return null
-    if (input.required && !value.trim()) {
-      window.alert(`${input.label} is required.`)
-      return null
-    }
-    if (value.trim()) values[input.key] = value.trim()
-  }
-  return values
-}
-
-function RunButton({ workflowId, inputs }: { workflowId: string; inputs: WorkflowDefinitionWire["inputs"] }) {
-  const navigate = useNavigate()
-  const queryClient = useQueryClient()
-  const start = useMutation({
-    mutationFn: (input: Record<string, string>) => api.startWorkflowRunV2(workflowId, input),
-    onSuccess: (detail) => {
-      queryClient.setQueryData(queryKeys.workflows.run(workflowId, detail.id), detail)
-      void queryClient.invalidateQueries({ queryKey: queryKeys.workflows.runs(workflowId) })
-      navigate(`/workflow/${encodeURIComponent(workflowId)}/runs/${encodeURIComponent(detail.id)}`)
-    },
-  })
-
-  return (
-    <button
-      type="button"
-      onClick={() => {
-        const input = collectRunInput(inputs)
-        if (input) start.mutate(input)
-      }}
-      disabled={start.isPending}
-      title={start.isError ? (start.error instanceof Error ? start.error.message : "Failed to start run.") : undefined}
-      className="inline-flex h-8 shrink-0 items-center gap-1.5 rounded-full bg-[var(--accent)] px-3.5 text-[length:var(--text-footnote)] font-[var(--weight-semibold)] text-[var(--accent-contrast)] transition-opacity hover:opacity-90 disabled:opacity-50" // jinn-shell: ok editor run control, not page chrome
-    >
-      <Play className="size-3.5" aria-hidden />
-      {start.isPending ? "Starting…" : "Run"}
-    </button>
   )
 }
